feat(ds): shard server workflow in two
