+++ svc/pkg/ds/src/lib.rs
@@ -11,10 +11,13 @@ pub fn registry() -> WorkflowResult<Registry> {
 
 	let mut registry = Registry::new();
 	registry.register_workflow::<server::Workflow>()?;
-	registry.register_workflow::<server::destroy::Workflow>()?;
-	registry.register_workflow::<server::nomad_alloc_plan::Workflow>()?;
-	registry.register_workflow::<server::nomad_alloc_update::Workflow>()?;
-	registry.register_workflow::<server::nomad_eval_update::Workflow>()?;
+	registry.register_workflow::<server::nomad::Workflow>()?;
+	registry.register_workflow::<server::nomad::destroy::Workflow>()?;
+	registry.register_workflow::<server::nomad::alloc_plan::Workflow>()?;
+	registry.register_workflow::<server::nomad::alloc_update::Workflow>()?;
+	registry.register_workflow::<server::nomad::eval_update::Workflow>()?;
+	registry.register_workflow::<server::nomad::eval_update::Workflow>()?;
+	// registry.register_workflow::<server::pegboard::Workflow>()?;
 
 	Ok(registry)
 }
